% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{bayescm_control}
\alias{bayescm_control}
\title{Controls Initialization of JAGS Model and MCMC Sampling}
\arguments{
\item{...}{Arguments passed to \code{\link[rjags]{jags.model}} and
\code{\link[rjags]{coda.samples}}.See arguments of both function.}
}
\value{
An object of class \code{bayescm_control} to use in
    \link{bayesgct} and \link{bayescmp}.
}
\description{
Parameters to control initialization of JAGS model,
    \code{\link[rjags]{jags.model}} and MCMC sampling,
    \code{\link[rjags]{jags.model}}.
}
\author{
Eduardo Jr
}
