% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{bayescmp}
\alias{bayescmp}
\title{Fitting COM-Poisson (mean parametrization) Model Based on MCMC}
\usage{
bayescmp(formula, data, sumto, .control_model = control_model(),
  .control_samples = control_samples())
}
\arguments{
\item{formula}{A formula to define fixed effects.}

\item{data}{The data frame.}

\item{sumto}{Upper bound to infinite sum of the normalization
constant}

\item{.control_model}{See \link{control_model}.}

\item{.control_samples}{See \link{control_samples}.}
}
\description{
Build matrices and organize data to use
    \code{\link[rjags]{jags.model}}
    and\code{\link[rjags]{coda.samples}} for analysis of Bayesian
    models using Markov Chain Monte Carlo (MCMC).
}
\examples{

\donttest{

# Simulate data
beta <- c(3, -1)
X <- cbind(1, runif(50))
y <- rpois(50, lambda = exp(X \%*\% beta))

# Sampling and summarise posterior
model <- bayescmp(y ~ X - 1, sumto = 100L)
vapply(model$samples, function(x) apply(x, 2L, mean), double(3))

}

}
\author{
Eduardo Jr
}
