% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_model}
\alias{control_model}
\title{rdname bayescm_control}
\usage{
control_model(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[rjags]{jags.model}}. See
the function arguments.}
}
\description{
rdname bayescm_control
}
