% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{control_samples}
\alias{control_samples}
\title{rdname bayescm_control}
\usage{
control_samples(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[rjags]{coda.samples}}. See
the function arguments.}
}
\description{
rdname bayescm_control
}
